chore: rename numbered test split files to describe their concern
I renamed the numbered test split files under tests/parse/ and
tests/engine_integration/ to names that describe the concern each file
actually holds, following the naming/subdivision standard, and updated the
mod declarations in the aggregators to match. This is a pure rename: the
test logic is unchanged and the suite compiles and runs with the same test
count as before.

I am deferring the internal/engine/ file-layout subdivision to a separate
change to avoid a large, risky churn in one commit.

Refs #553
