+++ tests/engine_integration/cli_lifecycle.rs
