+++ tests/parse/coverage.rs
@@ -1,9 +1,9 @@
 //! Parse tests for features present in the type system but not previously
 //! covered. Split into part files to keep each under the source line limit.
 
-#[path = "coverage/part1.rs"]
-mod part1;
-#[path = "coverage/part2.rs"]
-mod part2;
-#[path = "coverage/part3.rs"]
-mod part3;
+#[path = "coverage/deploy_build.rs"]
+mod deploy_build;
+#[path = "coverage/network_volume.rs"]
+mod network_volume;
+#[path = "coverage/resources_secrets.rs"]
+mod resources_secrets;
