+++ tests/parse/fields.rs
@@ -1,7 +1,7 @@
 //! Field-level parse tests. Split into part files to keep each under the
 //! source line limit.
 
-#[path = "fields/part1.rs"]
-mod part1;
-#[path = "fields/part2.rs"]
-mod part2;
+#[path = "fields/extends_build.rs"]
+mod extends_build;
+#[path = "fields/runtime_options.rs"]
+mod runtime_options;
